+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* Rename `createmodule` cmd to `startapp` to mimic Django
 * Fix the import issues that was causing [#37](https://github.com/shopyo/shopyo/issues/37). Now `shopyo new` command will not work but this is moved to [shopyo-factory](https://github.com/shopyo/shopyo-factory)
 * add changelog link to `setup.cfg` so that it shows on Shopyo PyPi page